requirements.txt
streamlit==1.34.0
numpy==1.24.4
Pillow==9.5.0
pytesseract==0.3.10
openai==0.27.0
opencv-python-headless==4.5.5.62
